% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adresse_api_koord.R
\encoding{UTF-8}
\name{adresse_api_koord}
\alias{adresse_api_koord}
\title{Adresser til koordinater}
\usage{
adresse_api_koord(postnummer, adresse, format = "sf", crs = 25833)
}
\arguments{
\item{postnummer}{Karaktervektor med postnummer.}

\item{adresse}{Karaktervektor med gateadresse (og gatenummer hvis tilgjengelig).}

\item{format}{Format på datasettet som returneres. Default er satt til "sf" (sf-objekt), men det er også mulig å sette format = "tibble" eller "data.frame".}

\item{crs}{Numerisk vektor med ønsket koordinatsystem for koordinatene. Default er satt til CRS 25833.}
}
\value{
Objekt med koordinatene til adressene.
}
\description{
Funksjon for å finne koordinater fra matrikkelen (adresse-API fra Kartverket).
}
\examples{
fra <- adresse_api_koord(postnummer = "0177",
                        adresse = "Akersveien 26")
}
