% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beregne_avstand.R
\encoding{UTF-8}
\name{beregne_avstand}
\alias{beregne_avstand}
\title{Beregne korteste kjøreavstand (igraph)}
\usage{
beregne_avstand(from_node, to_node, enhet = "FT_MINUTES", path = F)
}
\arguments{
\item{from_node}{Numerisk verdi med node-ID (for en vektor med flere node-ID-er, se funksjonen beregne_avstand_cpp).}

\item{to_node}{Numerisk verdi med node-ID (for en vektor med flere node-ID-er, se funksjonen beregne_avstand_cpp).}

\item{enhet}{Karaktervektor med "FT_MINUTES" for å få korteste kjørerute målt i minutter eller "LENGTH" for korteste kjørerute målt i meter.}

\item{path}{Boolsk. Dersom path = TRUE returneres nodelenken som den korteste kjøreruten består av.}
}
\value{
Vektor med korteste kjørerute i minutter eller meter. Dersom path = T returneres nodelenken som den korteste kjøreruten består av.
}
\description{
Funksjon for å beregne korteste kjøreavstand (i minutter eller meter) mellom noder i vegnettet. Funksjonen kan også returnere nodelenken (path) som den korteste kjøreruten består av.
For å bruke funksjonen må vegnettet ha blitt lastet inn og omgjort til et tbl_graph-objekt som heter "graph". Dette gjøres med funksjonen vegnett_to_R.
}
\examples{
avstand_min <- beregne_avstand(from_node = 26956,
                               to_node = 210373,
                               enhet = "FT_MINUTES")

avstand_meter <- beregne_avstand(from_node = 26956,
                                 to_node = 210373,
                                 enhet = "LENGTH")

path <- beregne_avstand(from_node = 26956,
                        to_node = 210373,
                        enhet = "FT_MINUTES",
                        path = T)
}
