% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beregne_avstand_cpp.R
\encoding{UTF-8}
\name{beregne_avstand_cpp}
\alias{beregne_avstand_cpp}
\title{Beregne korteste kjøreavstand (cppRouting)}
\usage{
beregne_avstand_cpp(
  from_node_ID,
  to_node_ID,
  enhet = "FT_MINUTES",
  dist = "all"
)
}
\arguments{
\item{from_node_ID}{Numerisk vektor med én eller flere node-ID-er.}

\item{to_node_ID}{Numerisk vektor med én eller flere node-ID-er.}

\item{enhet}{Karaktervektor med "FT_MINUTES" for å få korteste kjørerute målt i minutter eller "LENGTH" for korteste kjørerute målt i meter.}

\item{dist}{Karaktervektor som angir om man ønsker alle kjørerutene per from_node. Dersom man kun ønsker det korteste kjøreruten per from_node angir man dist = "min". "max" angir den lengste kjøreruten per from_node.}
}
\value{
Objekt (data.frame) med hvor mange minutter eller meter kjøreavstand det er mellom de angitte fra- og tilpunktene.
}
\description{
Funksjon for å beregne korteste kjøreavstand (i minutter eller meter) mellom to eller flere noder i vegnettet. Fungerer også med vektorer med flere adresser for start_node og/eller to_node.
For å bruke funksjonen må vegnettet ha blitt lastet inn og omgjort til et cppRouting-objekt som heter "graph_cppRouting_FT_MINUTES" eller "graph_cppRouting_LENGTH". Dette gjøres med funksjonen vegnett_to_R.
}
\examples{
avstand_cpp_min <- beregne_avstand_cpp(26956,
                                       210373,
                                       enhet = "FT_MINUTES")

avstand_cpp_meter <- beregne_avstand_cpp(26956,
                                         210373,
                                         enhet = "LENGTH")
}
