% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords_to_google.R
\encoding{UTF-8}
\name{coords_to_google}
\alias{coords_to_google}
\title{Convert coordinates to Google Maps}
\usage{
coords_to_google(coords, crs_out = 25833)
}
\arguments{
\item{coords}{An sf object with a geometry column that will be converted to CRS 4326.}

\item{crs_out}{Chosen coordinate reference system (CRS) for the geometry column of the returned sf object.}
}
\value{
An sf object with a new column added (\code{coords_google}), and an addtional sf geometry column with the chosen CRS.
}
\description{
\code{coords_to_google} is a function to convert coordinates of an sf object to a format that is easy to copy and paste into Google Maps.
}
\examples{
address_to_coords(zip_code = "0177", address = "Akersveien 26") \%>\%
  coords_to_google()
}
