% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata.R
\docType{data}
\name{graph_cppRouting_meters_sampledata}
\alias{graph_cppRouting_meters_sampledata}
\title{Sample data (graph_cppRouting_meters)}
\format{
A \code{cppRouting} object
}
\source{
\href{https://kartkatalog.geonorge.no/metadata/nvdb-ruteplan-nettverksdatasett/8d0f9066-34f9-4423-be12-8e8523089313/}{National roads database - road network for routing}
}
\usage{
data(graph_cppRouting_meters_sampledata)
}
\description{
Sample of the Norwegian Road Network (\code{vegnettRuteplan_FGDB_20210528.gdb}) downloaded from Geonorge. Only the first 1000 rows from Oslo county are included.
}
\examples{
data(graph_cppRouting_meters_sampledata)
class(graph_cppRouting_meters_sampledata)
}
\keyword{datasets}
