% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koords_to_google.R
\encoding{UTF-8}
\name{koords_to_google}
\alias{koords_to_google}
\title{Koordinater til Google Maps}
\usage{
koords_to_google(koords, crs_out = 25833)
}
\arguments{
\item{koords}{sf-objekt med en geometrikolonne som skal omgjøres til CRS 4326}

\item{crs_out}{Ønsket koordinatsystem (CRS) for geometrikolonnen til sf-objektet som returneres}
}
\value{
sf-objekt med en ny kolonne lagt til (coords_google)
}
\description{
Funksjon for å omgjøre koordinater til et format som er enkelt å klippe og lime inn i Google Maps.
}
\examples{
fra <- adresse_api_koord(postnummer = "0177",
                         adresse = "Akersveien 26") \%>\%
  koords_to_google()

}
