% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_koord.R
\encoding{UTF-8}
\name{node_koord}
\alias{node_koord}
\title{Knytte koordinater til noder i vegnettet}
\usage{
node_koord(
  koords,
  fra_til = "fra",
  ID_col = "ID",
  crs_out = 25833,
  knn = 1,
  membership = F
)
}
\arguments{
\item{koords}{sf-objekt med koordinatene man ønsker å plassere på vegnettet.}

\item{fra_til}{Karaktervektor med "fra" dersom noden skal være et startpunkt eller "til" dersom noden skal være et stoppunkt.}

\item{ID_col}{Karaktervektor med navnet på ID-kolonnen.}

\item{crs_out}{Numerisk vektor for ønsket koordinatsystem.}

\item{knn}{Numerisk vektor med antall noder per som blir returnert per koordinat. Dersom knn = 1 returneres den nærmeste noden til kooordinatene, dersom knn = 2 returneres de to nærmeste nodene til koordinatene osv.}

\item{membership}{Boolsk. Dersom TRUE begrenses søket etter noder til å kun gjelde noder som tilhører et tilknyttet vegnettverk til fra- eller tilpunktet (kun mulig for enten fra eller til).}
}
\value{
Objekt (data.frame) med kolonnene from_node/to_node, membership, coords_google_from_node/coords_google_to_node, nabor og ID
}
\description{
Funksjon for å finne nærmeste punkter (noder) i vegnettet til angitte koordinater.
}
\examples{
fra <- adresse_api_koord(postnummer = "0177",
                         adresse = "Akersveien 26")
from_node <- node_koord(koords = fra, fra_til = "fra")

til <- adresse_api_koord(postnummer = "2211",
                         adresse = "Otervegen 23")
to_node <- node_koord(koords = til, fra_til = "til")
}
