% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata.R
\docType{data}
\name{nodes_sampledata}
\alias{nodes_sampledata}
\title{Sample data (nodes)}
\format{
An object of class \code{sf}.
}
\source{
\href{https://kartkatalog.geonorge.no/metadata/nvdb-ruteplan-nettverksdatasett/8d0f9066-34f9-4423-be12-8e8523089313/}{National roads database - road network for routing}
}
\usage{
data(nodes_sampledata)
}
\description{
Sample of the Norwegian Road Network (\code{vegnettRuteplan_FGDB_20210528.gdb}) downloaded from Geonorge. Only the first 1000 rows from Oslo county are included.
}
\examples{
data(nodes_sampledata)
class(nodes_sampledata)
head(nodes_sampledata)
}
\keyword{datasets}
