% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_leaflet.R
\encoding{UTF-8}
\name{path_leaflet}
\alias{path_leaflet}
\title{Visualize the shortest path with Leaflet}
\usage{
path_leaflet(path, graph_object = graph)
}
\arguments{
\item{path}{Object (\code{list}) that has been created with the function \code{\link[=shortest_path_igraph]{shortest_path_igraph()}}, where \code{path = TRUE}.}

\item{graph_object}{The road network structured as a tidy graph (\code{tbl_graph} object). This can be done with the function \code{\link[=vegnett_to_R]{vegnett_to_R()}}.}
}
\value{
Interactive Leaflet map that shows the shortest path (in minutes or meters) between a chosen from and to node in the road network.
}
\description{
The function \code{path_leaflet} visualizes the shortest path (in minutes or meters) that has been calculated with the function \code{\link[=shortest_path_igraph]{shortest_path_igraph()}} (where \code{path = TRUE}).
Internet connection is required to load the background tiles.
}
\examples{
vegnett_list <- vegnett_to_R(vegnett = vegnett_sampledata,
                             year = 2021,
                             fromnodeID = "FROMNODEID",
                             tonodeID = "TONODEID",
                             FT_minutes = "FT_MINUTES",
                             TF_minutes = "TF_MINUTES",
                             meters = "SHAPE_LENGTH")

graph <- vegnett_list[[1]]
nodes <- vegnett_list[[2]]
edges <- vegnett_list[[3]]
graph_cppRouting_minutes <- vegnett_list[[4]]
graph_cppRouting_meters <- vegnett_list[[5]]

path <- shortest_path_igraph(from_node_ID = 25,
                             to_node_ID = 33,
                             unit = "minutes",
                             path = TRUE)

path_leaflet(path)

}
