% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortest_path_cppRouting.R
\encoding{UTF-8}
\name{shortest_path_cppRouting}
\alias{shortest_path_cppRouting}
\title{Shortest path (cppRouting)}
\usage{
shortest_path_cppRouting(
  from_node_ID,
  to_node_ID,
  unit = "minutes",
  dist = "all",
  graph_cppRouting_object = graph_cppRouting_minutes
)
}
\arguments{
\item{from_node_ID}{Numeric vector with one more from node ID’s.}

\item{to_node_ID}{Numeric vector with one more to node ID’s.}

\item{unit}{Character vector with "minutes" to calculate the shortest path in minutes or "meters" for the shortest path in meters.}

\item{dist}{Character vector that specifies if all the shortest paths between all the supplied from and to nodes are returned, or if only the minimum/maximum value per from node ID is returned.}

\item{graph_cppRouting_object}{The road network structured as a cppRouting graph object. This can be done with the function \code{\link[=vegnett_to_R]{vegnett_to_R()}}.}
}
\value{
Object (data.frame) with how many minutes or meters the shortest path is between the supplied from and to node ID’s.
}
\description{
The function \code{shortest_path_cppRouting} can be used to calculate the shortest path (either in minutes or meters) between two or more nodes in the Norwegian road network. The function also works with vectors with multiple from and to node ID’s.
Before the function can be used, the road network must be converted to a cppRouting object that is called \code{graph_cppRouting_minutes} or \code{graph_cppRouting_meters}. This can be done with the function \code{\link[=vegnett_to_R]{vegnett_to_R()}}.
}
\examples{
\dontrun{
distance_min <- shortest_path_cppRouting(26956,
                                       210373,
                                       unit = "minutes")

distance_meter <- shortest_path_cppRouting(26956,
                                         210373,
                                         unit = "meters")
                                         }
}
