% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegnett_sampledata.R
\docType{data}
\name{vegnett_sampledata}
\alias{vegnett_sampledata}
\title{Sample data}
\format{
An object of class sf.
}
\source{
\href{https://kartkatalog.geonorge.no/metadata/nvdb-ruteplan-nettverksdatasett/8d0f9066-34f9-4423-be12-8e8523089313/}{National roads database - road network for routing}
}
\usage{
data(vegnett_sampledata)
}
\description{
Sample of the Norwegian Road Network (NVDB Ruteplan Nettverksdatasett FGDB) downloaded from Geonorge on August 8 2022. Only the first 1000 rows from Oslo county are included.
}
\examples{
data(vegnett_sampledata)
class(vegnett_sampledata)
head(vegnett_sampledata)
}
\keyword{datasets}
