% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_to_coords.R
\encoding{UTF-8}
\name{address_to_coords}
\alias{address_to_coords}
\title{Convert addresses to coordinates}
\usage{
address_to_coords(zip_code, address, format = "sf", crs_out = 25833)
}
\arguments{
\item{zip_code}{Character vector with zip codes.}

\item{address}{Character vector with addresses (street name and house number).}

\item{format}{Format of the returned object. Default value is set to “sf” (which returns an sf object). It is also possible set the format to “tibble” or “data.frame”.}

\item{crs_out}{Numeric vector with the chosen coordinate reference system (CRS). Default value is set to CRS 25833.}
}
\value{
Object with coordinates to the supplied addresses.
}
\description{
Function to find coordinates to supplied Norwegian addresses. The function utilizes [the Norwegian Mapping Authority’s address API] (https://ws.geonorge.no/adresser/v1/).
}
\examples{
\dontrun{
from <- address_to_coords(zip_code = "0177",
                        address = "Akersveien 26")
                        }
}
