% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords_to_node.R
\encoding{UTF-8}
\name{coords_to_node}
\alias{coords_to_node}
\title{Connect coordinates to the nearest nodes in the road network}
\usage{
coords_to_node(
  coords,
  direction = "from",
  ID_col = "ID",
  crs_out = 25833,
  knn = 1,
  membership = F
)
}
\arguments{
\item{coords}{An sf object with the coordinates that should be connected to the road network.}

\item{direction}{Character vector with “from” if the points should be from nodes or “to” if the points should be to nodes.}

\item{ID_col}{Character vector with the name of the ID column. Default value is set to “ID”.}

\item{crs_out}{Numeric vector for the chosen coordinate reference system (CRS).}

\item{knn}{Numeric vector with the chosen number of nodes that should be returned for each of the coordinates. If \code{knn = 1} only the nearest nodes to the chosen coordinates will be returned. If \code{knn = 2} the two nearest nodes will be returned etc.}

\item{membership}{Logical. If TRUE the search for nodes is limited to nodes that belong to a road network that is connected either to the from or to nodes (only possible for either from or to). E.g. if you only want to search for from nodes that belong to the same road network as the to nodes, membership is set to FALSE in the search for to nodes and membership = TRUE for the from nodes (in that order).}
}
\value{
An object (data.frame) with the following columns; \code{from_nodeID}/\code{to_nodeID}, \code{membership_from_node}/\code{membership_to_node}, \code{coords_google_from_node}/\code{coords_google_to_node}, \code{knn_from_node}/\code{knn_to_node}, and \code{ID}.
}
\description{
Function to find the nearest nodes in the Norwegian road network for chosen coordinates.
}
\details{
Before the function can be used, the nodes of the road network must be converted to an sf object that is called \code{nodes}. This can be done with the function \code{\link[=vegnett_to_R]{vegnett_to_R()}}.
}
\examples{
\dontrun{
from <- address_to_coords(zip_code = "0177",
                         address = "Akersveien 26")
from_node <- coords_to_node(coords = from, direction = "from")

to <- address_to_coords(zip_code = "2211",
                         address = "Otervegen 23")
to_node <- coords_to_node(coords = to, direction = "to")
}
}
