% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_leaflet.R
\encoding{UTF-8}
\name{path_leaflet}
\alias{path_leaflet}
\title{Visualisere kjørerute med Leaflet}
\usage{
path_leaflet(path)
}
\arguments{
\item{path}{Objekt (list) som har blitt laget med funksjonen beregne_avstand (der path = TRUE).}
}
\value{
Interaktivt kart Leaflet som viser kjøreruten mellom et valgt fra- og tilpunkt på vegnettet.
}
\description{
Funksjonen path_leaflet visualiserer den korteste kjøreruten (i minutter eller meter) som har blitt beregnet med funksjonen beregne_avstand (der path = TRUE).
For å få lastet inn bakgrunnskart fra Leaflet kreves nettilgang.
}
\examples{
path_leaflet(path)
}
