% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_leaflet.R
\encoding{UTF-8}
\name{path_leaflet}
\alias{path_leaflet}
\title{Visualize the shortest path with Leaflet}
\usage{
path_leaflet(path)
}
\arguments{
\item{path}{Object (list) that has been created with the function \code{\link[=shortest_path_igraph]{shortest_path_igraph()}} (where path = TRUE).}
}
\value{
Interactive Leaflet map that shows the shortest path (in minutes or meters) between a chosen from and to node in the road network.
}
\description{
The function \code{path_leaflet} visualizes the shortest path (in minutes or meters) that has been calculated with the function \code{\link[=shortest_path_igraph]{shortest_path_igraph()}} (where path = TRUE).
Internet connection is required to load the background tiles.
}
\examples{
\dontrun{
path_leaflet(path)
}
}
