% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortest_path_igraph.R
\encoding{UTF-8}
\name{shortest_path_igraph}
\alias{shortest_path_igraph}
\title{Shortest path (igraph)}
\usage{
shortest_path_igraph(from_node_ID, to_node_ID, unit = "FT_MINUTES", path = F)
}
\arguments{
\item{from_node_ID}{Numeric value with the from node ID (if multiple node ID’s are to be used, see the function \code{shortest_path_cppRouting}).}

\item{to_node_ID}{Numeric value with the to node ID (if multiple node ID’s are to be used, see the function \code{shortest_path_cppRouting}).}

\item{unit}{Character vector with "FT_MINUTES" to calculate the shortest path in minutes or "LENGTH" for the shortest path in meters.}

\item{path}{Logical. If TRUE the node link with the shortest path is returned.}
}
\value{
Vector with the shortest path in minutes or meters. If path = TRUE the node link that the shortest path consists of is returned.
}
\description{
Function to calculate the shortest path (either in minutes or meters) between nodes in the Norwegian road network. The function can also return the node link (path) that the shortest path consists of.
}
\details{
Before the function can be used, the road network must be converted to a tbl_graph object that is called “graph”. This can be done with the function \code{vegnett_to_R }.
}
\examples{
\dontrun{
distance_min <- shortest_path_igraph(from_node_ID = 26956,
                               to_node_ID = 210373,
                               unit = "FT_MINUTES")

distance_meter <- shortest_path_igraph(from_node_ID = 26956,
                                 to_node_ID = 210373,
                                 unit = "LENGTH")

path <- shortest_path_igraph(from_node_ID = 26956,
                        to_node_ID = 210373,
                        unit = "FT_MINUTES",
                        path = T)
                        }
}
