% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata.R
\docType{data}
\name{vegnett_sampledata}
\alias{vegnett_sampledata}
\title{Sample data (vegnett)}
\format{
An object of class \code{sf}.
}
\source{
\href{https://kartkatalog.geonorge.no/metadata/nvdb-ruteplan-nettverksdatasett/8d0f9066-34f9-4423-be12-8e8523089313/}{National roads database - road network for routing}
}
\usage{
data(vegnett_sampledata)
}
\description{
Sample of the Norwegian Road Network (\code{vegnettRuteplan_FGDB_20210528.gdb}) downloaded from Geonorge. Only the first 1000 rows from Oslo county are included.
}
\examples{
data(vegnett_sampledata)
class(vegnett_sampledata)
head(vegnett_sampledata)
}
\keyword{datasets}
