% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegnett_to_R.R
\encoding{UTF-8}
\name{vegnett_to_R}
\alias{vegnett_to_R}
\title{Konverte vegnettet til nettverksgrafer i R}
\usage{
vegnett_to_R(vegnett, crs_out = 25833)
}
\arguments{
\item{vegnett}{Vegnettet som et sf-objekt.}

\item{crs_out}{Numerisk vektor for ønsket koordinatsystem (CRS).}
}
\value{
Liste med følgende objekter:

[1] graph: vegnettet strukturert som "tidy graph" (tbl_graph-objekt)

[2] nodes: nodene til vegnettet (sf-objekt)

[3] edges: veglenkene som vegnettet består av (data.frame)

[4] graph_cppRouting_FT_MINUTES: vegnettet strukturert som "cppRouting graph" med kostnaden per veglenke i minutter (cppRouting-objekt)

[5] graph_cppRouting_LENGTH: vegnettet strukturert som "cppRouting graph" med kostnaden per veglenke i meter (cppRouting-objekt)
}
\description{
Funksjon for å omgjøre vegnettet fra Nasjonal vegdatabank (NVDB) til formater tilpasset nettverksanalyse i R (tbl_graph og cppRouting)
}
\examples{
vegnett_list <- vegnett_to_R(vegnett = vegnett)

graph <- vegnett_list[[1]]
nodes <- vegnett_list[[2]]
edges <- vegnett_list[[3]]
graph_cppRouting_FT_MINUTES <- vegnett_list[[4]]
graph_cppRouting_LENGTH <- vegnett_list[[5]]
}
